requirements-base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements-base.in
#
anyio==3.6.2
    # via starlette
fastapi==0.91.0
    # via -r requirements-base.in
geographiclib==2.0
    # via geopy
geopy==2.3.0
    # via -r requirements-base.in
h11==0.14.0
    # via
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
hypercorn==0.14.3
    # via -r requirements-base.in
hyperframe==6.0.1
    # via h2
idna==3.4
    # via anyio
priority==2.0.0
    # via hypercorn
pydantic==1.10.4
    # via
    #   -r requirements-base.in
    #   fastapi
python-dotenv==0.21.1
    # via -r requirements-base.in
sniffio==1.3.0
    # via anyio
starlette==0.24.0
    # via fastapi
toml==0.10.2
    # via hypercorn
typing-extensions==4.4.0
    # via pydantic
wsproto==1.2.0
    # via hypercorn
